+++ blockchain/app.py
@@ -2,149 +2,12 @@ import argparse
 import uuid
 import logging
 import sys
-from typing import TYPE_CHECKING
-from urllib.parse import urlparse, urljoin
+from urllib.parse import urlparse
 
-from flasgger import Swagger
-from flask import Flask, jsonify, request
-
-from blockchain import __meta__, __title__
+from blockchain.api import APP
 from blockchain.database import DB_TYPES
-from blockchain.impl import Blockchain
-from blockchain.utils import asbool, get_logger
-
-if TYPE_CHECKING:
-    from blockchain.database import Database
-
-
-class BlockchainWebApp(Flask):
-    blockchain = None   # type: Blockchain
-    node = None         # type: str
-    db = None           # type: Database
-
-
-# Instantiate the Node
-app = BlockchainWebApp(__name__)
-Swagger(
-    app,
-    merge=True, config={"openapi": "3.0.2", "specs_route": "/api"},
-    template={
-        "info": {
-            "title": __title__,
-            "description": __meta__["Summary"],
-            "contact": {
-                "responsibleOrganization": __meta__["Author"],
-                "responsibleDeveloper": __meta__["Maintainer"],
-                "email": __meta__["Maintainer-email"],
-                "url": __meta__["Home-page"],  # url=... in setup
-            },
-            # "termsOfService": "http://me.com/terms",
-            "license": __meta__["License"],
-            "version": __meta__["Version"]
-        }
-    }
-)
-
-
-@app.route("/")
-def frontpage():
-    body = {
-        "description": "Blockchain Node",
-        "node": app.node,
-        "links": [
-            {"rel": "api", "href": urljoin(request.url, "/api")}
-        ]
-    }
-    return jsonify(body), 200
-
-
-@app.route("/mine", methods=["GET"])
-def mine():
-    # We run the proof of work algorithm to get the next proof...
-    last_block = app.blockchain.last_block
-    proof = app.blockchain.proof_of_work(last_block)
-
-    # We must receive a reward for finding the proof.
-    # The sender is "0" to signify that this node has mined a new coin.
-    app.blockchain.new_transaction(
-        sender="0",
-        recipient=app.node,
-        amount=1,
-    )
-
-    # Forge the new Block by adding it to the chain
-    previous_hash = app.blockchain.hash(last_block)
-    block = app.blockchain.new_block(proof, previous_hash)
-
-    response = {
-        "message": "New Block Forged",
-        "index": block["index"],
-        "transactions": block["transactions"],
-        "proof": block["proof"],
-        "previous_hash": block["previous_hash"],
-    }
-    return jsonify(response), 200
-
-
-@app.route("/transactions/new", methods=["POST"])
-def new_transaction():
-    values = request.get_json()
-
-    # Check that the required fields are in the POST'ed data
-    required = {"sender": str, "recipient": str, "amount": int}
-    if not all(k in values and isinstance(values[k], required[k]) for k in required):
-        return "Missing values amongst: {}".format(required), 400
-
-    # Create a new Transaction
-    index = app.blockchain.new_transaction(values["sender"], values["recipient"], values["amount"])
-
-    response = {"message": f"Transaction will be added to Block {index}"}
-    return jsonify(response), 201
-
-
-@app.route("/chain", methods=["GET"])
-def full_chain():
-    detail = asbool(request.args.get("detail"))
-    response = {
-        "chain": app.blockchain.json(detail=detail),
-        "length": len(app.blockchain.blocks),
-    }
-    return jsonify(response), 200
-
-
-@app.route("/nodes", methods=["GET"])
-def list_nodes():
-    return jsonify({"nodes": list(app.blockchain.nodes)}), 200
-
-
-@app.route("/nodes/register", methods=["POST"])
-def register_nodes():
-    values = request.get_json()
-
-    nodes = values.get("nodes")
-    if nodes is None:
-        return "Error: Please supply a valid list of nodes", 400
-
-    for node in nodes:
-        app.blockchain.register_node(node)
-
-    response = {
-        "message": "New nodes have been added",
-        "total_nodes": list(app.blockchain.nodes),
-    }
-    return jsonify(response), 201
-
-
-@app.route("/nodes/resolve", methods=["GET"])
-def consensus():
-    replaced = app.blockchain.resolve_conflicts()
-
-    if replaced:
-        response = {"message": "Our chain was replaced", "new_chain": app.blockchain.chain}
-    else:
-        response = {"message": "Our chain is authoritative", "chain": app.blockchain.chain}
-    app.db.save_chain(app.blockchain)
-    return jsonify(response), 200
+from blockchain.impl import Blockchain, Node
+from blockchain.utils import get_logger
 
 
 class DatabaseTypeAction(argparse.Action):
@@ -159,12 +22,12 @@ class DatabaseTypeAction(argparse.Action):
             db_conn = uri.netloc
         else:
             if not uri.path.startswith("/"):
-                raise ValueError("Database URI without scheme must be an absolute path: [{}]".format(uri))
+                raise ValueError(f"Database URI without scheme must be an absolute path: [{uri!s}]")
             db_type = "file"
             db_conn = uri.path
         db_impl = DB_TYPES.get(db_type)
         if not db_impl:
-            raise ValueError("Unknown database type: [{}]".format(db_type))
+            raise ValueError(f"Unknown database type: [{db_type!s}]")
         setattr(namespace, self.dest, db_impl(db_conn))
 
 
@@ -172,30 +35,51 @@ def main():
     parser = argparse.ArgumentParser(prog="blockchain", description="Blockchain Node Web Application")
     parser.add_argument("-p", "--port", default=5000, type=int, help="port to listen on")
     parser.add_argument("-n", "--node", help="Unique identifier of the node. Generate one if omitted.")
-    parser.add_argument("--db", "--database", default="file", action=DatabaseTypeAction,
-                        help="Database to use. Formatted as [type://connection-detail].")
+    parser.add_argument("-N", "--nodes", nargs="*", action="append", type=str,
+                        help="Node endpoints the blockchains should resolve consensus against.")
+
+    db_args = parser.add_argument_group(title="Database", description="Database options.")
+    db_args.add_argument("--db", "--database", required=True, action=DatabaseTypeAction,
+                         help="Database to use. Formatted as [type://connection-detail].")
+    db_args.add_argument("--new", action="store_true", help="Generate the new blockchain with genesis block.")
 
     log_args = parser.add_argument_group(title="Logger", description="Logging control.")
     level_args = log_args.add_mutually_exclusive_group()
     level_args.add_argument("-q", "--quiet", action="store_true", help="Disable logging except errors.")
     level_args.add_argument("-d", "--debug", action="store_true",
                             help="Debug level logging. This also enables error traceback outputs in responses.")
+    log_args.add_argument("-v", "--verbose", action="store_true", help="Enforce verbose logging to stdout.")
+    log_args.add_argument("-l", "--log", help="output file to write generated logs.")
     args = parser.parse_args()
 
     # set full module config
     level = logging.DEBUG if args.debug else logging.ERROR if args.quiet else logging.ERROR
-    logger = get_logger("blockchain", level=level)
+    logger = get_logger("blockchain", level=level, force_stdout=args.verbose, file=args.log)
     if level == logging.DEBUG:
-        app.debug = True
+        APP.debug = True
 
     try:
+        host = "0.0.0.0"
         port = args.port
-        app.db = args.db
+        APP.url = f"{host}:{port}"
+        APP.db = args.db
+
+        if args.new:
+            chain = Blockchain()
+            logger.info("New blockchain: [%s]", chain.id)
+            APP.db.save_chain(chain)
+            sys.exit(0)
 
         # Generate a globally unique address for this node
-        app.node = args.node if args.node else str(uuid.uuid4()).replace("-", "")
-        app.blockchain = app.db.load_chain()
-        app.run(host="0.0.0.0", port=port)
+        APP.node = args.node if args.node else str(uuid.uuid4())
+        APP.blockchains = APP.db.load_multi_chain()
+        if args.nodes:
+            nodes = {node for node_list in args.nodes for node in node_list}  # flatten repeated -N, multi-URI per -N
+            APP.nodes = list(sorted([Node(node) for node in nodes], key=lambda n: n.url))
+            for node in APP.nodes:
+                if urlparse(node.url) == urlparse(APP.url):
+                    raise ValueError("Cannot use current APP endpoint as other consensus node endpoint.")
+        APP.run(host=host, port=port)
     except Exception as exc:
         logger.error("Unhandled error: %s", exc, exc_info=exc)
         sys.exit(-1)